cicd: only use forge

+++ test/account/unit-tests/Allowance.t.sol
@@ -12,13 +12,12 @@ import "../../../src/Account.sol";
 import {AccountTestBase} from "./AccountTestBase.sol";
 
 contract UNIT_Allowances is Test, AccountTestBase {
-  
   function setUp() public {
     setUpAccounts();
   }
 
   function testCanTransferWithoutPositiveAllowance() public {
-    int256 amount = 1e18;
+    int amount = 1e18;
     vm.startPrank(alice);
     transferToken(aliceAcc, bobAcc, usdcAsset, 0, amount);
     vm.stopPrank();
@@ -27,8 +26,9 @@ contract UNIT_Allowances is Test, AccountTestBase {
   function testCannotSetAllowanceFromNonOwner() public {
     AccountStructs.AssetAllowance[] memory assetAllowances = new AccountStructs.AssetAllowance[](0);
     vm.expectRevert(
-      abi.encodeWithSelector(Account.NotOwnerOrERC721Approved.selector, 
-        address(account), 
+      abi.encodeWithSelector(
+        Account.NotOwnerOrERC721Approved.selector,
+        address(account),
         alice,
         bobAcc,
         bob,
@@ -41,15 +41,10 @@ contract UNIT_Allowances is Test, AccountTestBase {
   }
 
   function testCannotTransferNegativeAllowanceWithoutAllowance() public {
-    int256 amount = 1e18;
+    int amount = 1e18;
     vm.expectRevert(
-      abi.encodeWithSelector(Allowances.NotEnoughSubIdOrAssetAllowances.selector, 
-        address(account), 
-        alice,
-        bobAcc,
-        -1e18,
-        0,
-        0
+      abi.encodeWithSelector(
+        Allowances.NotEnoughSubIdOrAssetAllowances.selector, address(account), alice, bobAcc, -1e18, 0, 0
       )
     );
     vm.startPrank(alice);
@@ -57,17 +52,12 @@ contract UNIT_Allowances is Test, AccountTestBase {
     vm.stopPrank();
   }
 
-  function testCannotTradeWithoutAllowance() public {    
+  function testCannotTradeWithoutAllowance() public {
     // expect revert
     vm.startPrank(alice);
     vm.expectRevert(
-      abi.encodeWithSelector(Allowances.NotEnoughSubIdOrAssetAllowances.selector, 
-        address(account), 
-        alice,
-        bobAcc,
-        -1e18,
-        0,
-        0
+      abi.encodeWithSelector(
+        Allowances.NotEnoughSubIdOrAssetAllowances.selector, address(account), alice, bobAcc, -1e18, 0, 0
       )
     );
     tradeTokens(aliceAcc, bobAcc, address(usdcAsset), address(coolAsset), 1e18, 1e18, 0, tokenSubId);
@@ -80,11 +70,7 @@ contract UNIT_Allowances is Test, AccountTestBase {
     vm.startPrank(bob);
     // bob allow alice to move its cool token
     AccountStructs.AssetAllowance[] memory assetAllowances = new AccountStructs.AssetAllowance[](1);
-    assetAllowances[0] = AccountStructs.AssetAllowance({
-      asset: coolAsset,
-      positive: 0,
-      negative: tradeAmount
-    });
+    assetAllowances[0] = AccountStructs.AssetAllowance({asset: coolAsset, positive: 0, negative: tradeAmount});
     account.setAssetAllowances(bobAcc, alice, assetAllowances);
     vm.stopPrank();
 
@@ -94,7 +80,7 @@ contract UNIT_Allowances is Test, AccountTestBase {
     vm.stopPrank();
 
     // test end state
-    uint256 usdcAllowanceLeft = account.positiveAssetAllowance(bobAcc, bob, usdcAsset, alice);
+    uint usdcAllowanceLeft = account.positiveAssetAllowance(bobAcc, bob, usdcAsset, alice);
     assertEq(usdcAllowanceLeft, 0);
   }
 
@@ -104,12 +90,8 @@ contract UNIT_Allowances is Test, AccountTestBase {
     vm.startPrank(bob);
     // bob allow alice to move its cool token
     AccountStructs.SubIdAllowance[] memory tokenSubIdAllowances = new AccountStructs.SubIdAllowance[](1);
-    tokenSubIdAllowances[0] = AccountStructs.SubIdAllowance({
-      asset: coolAsset,
-      subId: tokenSubId,
-      positive: 0,
-      negative: tradeAmount
-    });
+    tokenSubIdAllowances[0] =
+      AccountStructs.SubIdAllowance({asset: coolAsset, subId: tokenSubId, positive: 0, negative: tradeAmount});
     account.setSubIdAllowances(bobAcc, alice, tokenSubIdAllowances);
     vm.stopPrank();
 
@@ -119,7 +101,7 @@ contract UNIT_Allowances is Test, AccountTestBase {
     vm.stopPrank();
 
     // test end state
-    uint256 tokenAllowanceLeft = account.negativeSubIdAllowance(bobAcc, bob, coolAsset, tokenSubId, alice);
+    uint tokenAllowanceLeft = account.negativeSubIdAllowance(bobAcc, bob, coolAsset, tokenSubId, alice);
     assertEq(tokenAllowanceLeft, 0);
   }
 
@@ -129,20 +111,12 @@ contract UNIT_Allowances is Test, AccountTestBase {
     vm.startPrank(bob);
     // bob allow alice to move its cool token both on "asset" and "subIdAllowance"
     AccountStructs.AssetAllowance[] memory assetAllowances = new AccountStructs.AssetAllowance[](1);
-    assetAllowances[0] = AccountStructs.AssetAllowance({
-      asset: coolAsset,
-      positive: 0,
-      negative: tradeAmount / 2
-    });
+    assetAllowances[0] = AccountStructs.AssetAllowance({asset: coolAsset, positive: 0, negative: tradeAmount / 2});
     account.setAssetAllowances(bobAcc, alice, assetAllowances);
 
     AccountStructs.SubIdAllowance[] memory tokenSubIdAllowances = new AccountStructs.SubIdAllowance[](1);
-    tokenSubIdAllowances[0] = AccountStructs.SubIdAllowance({
-      asset: coolAsset,
-      subId: tokenSubId,
-      positive: 0,
-      negative: tradeAmount / 2
-    });
+    tokenSubIdAllowances[0] =
+      AccountStructs.SubIdAllowance({asset: coolAsset, subId: tokenSubId, positive: 0, negative: tradeAmount / 2});
     account.setSubIdAllowances(bobAcc, alice, tokenSubIdAllowances);
     vm.stopPrank();
 
@@ -156,38 +130,30 @@ contract UNIT_Allowances is Test, AccountTestBase {
     assertEq(account.negativeAssetAllowance(bobAcc, bob, coolAsset, alice), 0);
   }
 
-  function testCannotTradeWithInsufficientTotalAllowance() public {    
-
+  function testCannotTradeWithInsufficientTotalAllowance() public {
     vm.startPrank(bob);
     // bob allow alice to move its cool token
     AccountStructs.AssetAllowance[] memory assetAllowances = new AccountStructs.AssetAllowance[](1);
-    assetAllowances[0] = AccountStructs.AssetAllowance({
-      asset: coolAsset,
-      positive: 0,
-      negative: 5e17
-    });
+    assetAllowances[0] = AccountStructs.AssetAllowance({asset: coolAsset, positive: 0, negative: 5e17});
     account.setAssetAllowances(bobAcc, alice, assetAllowances);
 
     AccountStructs.SubIdAllowance[] memory tokenSubIdAllowances = new AccountStructs.SubIdAllowance[](1);
-    tokenSubIdAllowances[0] = AccountStructs.SubIdAllowance({
-      asset: coolAsset,
-      subId: tokenSubId,
-      positive: 0,
-      negative: 4e17
-    });
+    tokenSubIdAllowances[0] =
+      AccountStructs.SubIdAllowance({asset: coolAsset, subId: tokenSubId, positive: 0, negative: 4e17});
     account.setSubIdAllowances(bobAcc, alice, tokenSubIdAllowances);
     vm.stopPrank();
 
     // expect revert
     vm.startPrank(alice);
     vm.expectRevert(
-      abi.encodeWithSelector(Allowances.NotEnoughSubIdOrAssetAllowances.selector, 
-        address(account), 
+      abi.encodeWithSelector(
+        Allowances.NotEnoughSubIdOrAssetAllowances.selector,
+        address(account),
         alice,
         bobAcc,
         -1e18,
         4e17, // tokenSubId allowance
-        5e17  // asset allowance
+        5e17 // asset allowance
       )
     );
 
@@ -196,7 +162,7 @@ contract UNIT_Allowances is Test, AccountTestBase {
     vm.stopPrank();
   }
 
-  function testCannotTradeWithInsufficientPositiveAllowance() public {    
+  function testCannotTradeWithInsufficientPositiveAllowance() public {
     uint tradeAmount = 1e18;
 
     // set that we need positive allowance to receive USDC
@@ -205,24 +171,21 @@ contract UNIT_Allowances is Test, AccountTestBase {
     vm.startPrank(bob);
     // bob allow alice to move its cool token
     AccountStructs.AssetAllowance[] memory assetAllowances = new AccountStructs.AssetAllowance[](1);
-    assetAllowances[0] = AccountStructs.AssetAllowance({
-      asset: coolAsset,
-      positive: 0,
-      negative: tradeAmount
-    });
+    assetAllowances[0] = AccountStructs.AssetAllowance({asset: coolAsset, positive: 0, negative: tradeAmount});
     account.setAssetAllowances(bobAcc, alice, assetAllowances);
     vm.stopPrank();
 
     // expect revert
     vm.startPrank(alice);
     vm.expectRevert(
-      abi.encodeWithSelector(Allowances.NotEnoughSubIdOrAssetAllowances.selector, 
-        address(account), 
+      abi.encodeWithSelector(
+        Allowances.NotEnoughSubIdOrAssetAllowances.selector,
+        address(account),
         alice,
         bobAcc,
         tradeAmount, // cannot increase amount!
         0, // tokenSubId allowance
-        0  // asset allowance
+        0 // asset allowance
       )
     );
 
@@ -231,21 +194,15 @@ contract UNIT_Allowances is Test, AccountTestBase {
     vm.stopPrank();
   }
 
-  function test3rdPartyAllowance() public {    
+  function test3rdPartyAllowance() public {
     address orderbook = address(0xb00c);
 
     // give orderbook allowance over both
     AccountStructs.AssetAllowance[] memory assetAllowances = new AccountStructs.AssetAllowance[](2);
-    assetAllowances[0] = AccountStructs.AssetAllowance({
-      asset: coolAsset,
-      positive: type(uint).max,
-      negative: type(uint).max
-    });
-    assetAllowances[1] = AccountStructs.AssetAllowance({
-      asset: usdcAsset,
-      positive: type(uint).max,
-      negative: type(uint).max
-    });
+    assetAllowances[0] =
+      AccountStructs.AssetAllowance({asset: coolAsset, positive: type(uint).max, negative: type(uint).max});
+    assetAllowances[1] =
+      AccountStructs.AssetAllowance({asset: usdcAsset, positive: type(uint).max, negative: type(uint).max});
 
     vm.startPrank(bob);
     account.setAssetAllowances(bobAcc, orderbook, assetAllowances);
@@ -268,21 +225,15 @@ contract UNIT_Allowances is Test, AccountTestBase {
     debtAsset.setNeedPositiveAllowance(true); // need allowance to increase balance
 
     // can transfer negative amount
-    int256 amount = 1e18;
-    
+    int amount = 1e18;
+
     vm.startPrank(alice);
     transferToken(aliceAcc, bobAcc, debtAsset, 0, -amount);
-    
 
     // cannot transfer positive amount
     vm.expectRevert(
-      abi.encodeWithSelector(Allowances.NotEnoughSubIdOrAssetAllowances.selector, 
-        address(account), 
-        alice,
-        bobAcc,
-        1e18,
-        0,
-        0
+      abi.encodeWithSelector(
+        Allowances.NotEnoughSubIdOrAssetAllowances.selector, address(account), alice, bobAcc, 1e18, 0, 0
       )
     );
     transferToken(aliceAcc, bobAcc, debtAsset, 0, amount);
@@ -290,11 +241,7 @@ contract UNIT_Allowances is Test, AccountTestBase {
 
     // can transfer positive amount after setting positive allowance
     AccountStructs.AssetAllowance[] memory assetAllowances = new AccountStructs.AssetAllowance[](1);
-    assetAllowances[0] = AccountStructs.AssetAllowance({
-      asset: debtAsset,
-      positive: uint(amount),
-      negative: 0
-    });
+    assetAllowances[0] = AccountStructs.AssetAllowance({asset: debtAsset, positive: uint(amount), negative: 0});
     vm.prank(bob);
     account.setAssetAllowances(bobAcc, alice, assetAllowances);
 
@@ -309,36 +256,27 @@ contract UNIT_Allowances is Test, AccountTestBase {
 
     vm.startPrank(bob);
     AccountStructs.AssetAllowance[] memory assetAllowances = new AccountStructs.AssetAllowance[](1);
-    assetAllowances[0] = AccountStructs.AssetAllowance({
-      asset: coolAsset,
-      positive: 0,
-      negative: tradeAmount
-    });
+    assetAllowances[0] = AccountStructs.AssetAllowance({asset: coolAsset, positive: 0, negative: tradeAmount});
     account.setAssetAllowances(bobAcc, alice, assetAllowances);
-    account.transferFrom(bob, charlie, bobAcc);  // transfer account to charlie
+    account.transferFrom(bob, charlie, bobAcc); // transfer account to charlie
     vm.stopPrank();
 
     uint charlieAcc = bobAcc;
 
-    // should revert because allowance is not transferred 
+    // should revert because allowance is not transferred
     vm.startPrank(alice);
     vm.expectRevert(
-      abi.encodeWithSelector(Allowances.NotEnoughSubIdOrAssetAllowances.selector, 
-        address(account), 
-        alice,
-        charlieAcc,
-        -1e18,
-        0,
-        0
+      abi.encodeWithSelector(
+        Allowances.NotEnoughSubIdOrAssetAllowances.selector, address(account), alice, charlieAcc, -1e18, 0, 0
       )
     );
     tradeTokens(aliceAcc, charlieAcc, address(usdcAsset), address(coolAsset), tradeAmount, tradeAmount, 0, tokenSubId);
     vm.stopPrank();
   }
 
-  function testERC721Approval() public {    
+  function testERC721Approval() public {
     uint tradeAmount = 1e18;
-    
+
     vm.startPrank(bob);
     account.approve(alice, bobAcc);
     vm.stopPrank();
@@ -350,17 +288,11 @@ contract UNIT_Allowances is Test, AccountTestBase {
 
     // revert with new account
     uint bobNewAcc = account.createAccount(bob, dumbManager);
-    mintAndDeposit(
-      bob,
-      bobNewAcc,
-      coolToken,
-      coolAsset,
-      tokenSubId,
-      tradeAmount
-    );
+    mintAndDeposit(bob, bobNewAcc, coolToken, coolAsset, tokenSubId, tradeAmount);
     vm.startPrank(alice);
     vm.expectRevert(
-      abi.encodeWithSelector(Allowances.NotEnoughSubIdOrAssetAllowances.selector,
+      abi.encodeWithSelector(
+        Allowances.NotEnoughSubIdOrAssetAllowances.selector,
         address(account),
         address(alice),
         bobNewAcc,
@@ -373,7 +305,7 @@ contract UNIT_Allowances is Test, AccountTestBase {
     vm.stopPrank();
   }
 
-  function testERC721ApprovalForAll() public {    
+  function testERC721ApprovalForAll() public {
     uint tradeAmount = 1e18;
 
     vm.startPrank(bob);
@@ -387,45 +319,31 @@ contract UNIT_Allowances is Test, AccountTestBase {
 
     // successful trade even with new account from same user
     uint bobNewAcc = account.createAccount(bob, dumbManager);
-    mintAndDeposit(
-      bob,
-      bobNewAcc,
-      coolToken,
-      coolAsset,
-      tokenSubId,
-      tradeAmount
-    );    
+    mintAndDeposit(bob, bobNewAcc, coolToken, coolAsset, tokenSubId, tradeAmount);
 
     vm.startPrank(alice);
     tradeTokens(aliceAcc, bobNewAcc, address(usdcAsset), address(coolAsset), tradeAmount, tradeAmount, 0, tokenSubId);
     vm.stopPrank();
   }
 
-  function testManagerInitiatedTransfer() public {    
+  function testManagerInitiatedTransfer() public {
     // successful trade without allowances
     vm.startPrank(address(dumbManager));
     tradeTokens(aliceAcc, bobAcc, address(usdcAsset), address(coolAsset), 1e18, 1e18, 0, tokenSubId);
     vm.stopPrank();
   }
 
-  function testAutoAllowanceWithNewAccount() public {    
+  function testAutoAllowanceWithNewAccount() public {
     // new user account with spender allowance
     vm.startPrank(alice);
     address user = vm.addr(100);
     uint userAcc = account.createAccountWithApproval(user, bob, dumbManager);
     vm.stopPrank();
-    mintAndDeposit(
-        user,
-        userAcc,
-        usdc,
-        usdcAsset,
-        0,
-        1e18
-    );
+    mintAndDeposit(user, userAcc, usdc, usdcAsset, 0, 1e18);
 
     // successful trade without allowances
     vm.startPrank(bob);
     tradeTokens(userAcc, bobAcc, address(usdcAsset), address(coolAsset), 1e18, 1e18, 0, tokenSubId);
     vm.stopPrank();
-  }  
+  }
 }
